Add handlers for creating new items
This commit adds the GET and POST methods for `/item/new`.

There's a struct for holding partial todo-item data (in case creation
fails and we want to retain entered information) but for now, creating
an item can't fail except at the repo level, which is probably not worth
handling yet.