Store timestamps as integers
This commit changes the storage format of timstamps to Unix timestamps
instead of RFC3... timestamps. The reason is because it's Python (and
likely other languages) don't support nanosecond timestamps (which is
the Chrono default) and would require a bunch of logic to interact with
`chrono::Local`. Storing everything in UTC erases some historical data
but **greatly** simplifies automation and integration with other
languages.